types package created and parse package updated to use it. Moved magefile template to its own file
